refactor Hotel module

+++ hotel-service/src/main/java/org/microarchitecturovisco/hotelservice/domain/Location.java
@@ -12,14 +12,12 @@ import lombok.*;
 public class Location {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String country;
 
     private String region;
 
-
     @OneToOne(mappedBy = "location")
     private Hotel hotel;
-
 }
